Allow pytest --pyargs jwst to work (#5319)
* Allow pytest --pyargs jwst to work

* Add pyargs to Travis job

* Remove MANIFEST.in and fix package_data glob patterns

* move set_velocity_aberration to jwst.lib
