+++ .travis.yml
@@ -41,6 +41,13 @@ jobs:
         # Github master until it is released to PyPI
         - pip install git+https://github.com/spacetelescope/stdatamodels
 
+    - name: Installed package with --pyargs
+      before_script: cd /
+      install:
+        - pip install .[test]
+      script:
+        - pytest --pyargs jwst
+
     - name: SDP dependencies in requirements-sdp.txt
       env:
         - CRDS_CONTEXT=jwst-edit
