+++ jwst/lib/tests/test_velocity_aberration.py
@@ -1,16 +1,12 @@
 """
 Test script for set_velocity_aberration.py
 """
+import subprocess
+
 from numpy import isclose
-import os
-import sys
+from astropy.io import fits
 
-sys.path.insert(
-    0,
-    os.path.join(os.path.dirname(__file__), '../../scripts')
-)
-
-import set_velocity_aberration as sva # noqa: E402
+import jwst.lib.set_velocity_aberration as sva
 
 # Testing constants
 GOOD_VELOCITY = (100.0, 100.0, 100.0)
@@ -57,3 +53,24 @@ def test_offset_zero_velocity():
     )
     assert isclose(delta_x, ZERO_OFFSET_X)
     assert isclose(delta_y, ZERO_OFFSET_Y)
+
+
+def test_velocity_aberration_script(tmpdir):
+    """Test the whole script on a FITS file"""
+    path = str(tmpdir.join("velocity_aberration_tmpfile.fits"))
+    hdulist = fits.HDUList(fits.PrimaryHDU())
+    hdulist.append(fits.ImageHDU(name="SCI"))
+    hdulist["PRIMARY"].header.append(('JWST_DX', GOOD_VELOCITY[0]))
+    hdulist["PRIMARY"].header.append(('JWST_DY', GOOD_VELOCITY[1]))
+    hdulist["PRIMARY"].header.append(('JWST_DZ', GOOD_VELOCITY[2]))
+    hdulist["SCI"].header.append(('RA_REF', GOOD_POS[0]))
+    hdulist["SCI"].header.append(('DEC_REF', GOOD_POS[1]))
+    hdulist.writeto(path)
+    hdulist.close()
+
+    subprocess.check_call(["set_velocity_aberration.py", path])
+
+    with fits.open(path) as hdulist:
+        assert isclose(hdulist["PRIMARY"].header['DVA_RA'], GOOD_OFFSET_X)
+        assert isclose(hdulist["PRIMARY"].header['DVA_DEC'], GOOD_OFFSET_Y)
+        assert isclose(hdulist["SCI"].header["VA_SCALE"], GOOD_SCALE_FACTOR)
