--- jwst/stpipe/tests/steps/log.cfg
@@ -1,1 +1,0 @@
-[*]
