+++ scripts/set_telescope_pointing.py
