+++ setup.cfg
@@ -2,6 +2,7 @@
 name = jwst
 description = Library for calibration of science observations from the James Webb Space Telescope
 long_description = Library for calibration of science observations from the James Webb Space Telescope
+long_description_content_type = text
 author = JWST calibration pipeline developers
 license = BSD-3-Clause
 url = https://github.com/spacetelescope/jwst
